Puts unit tests into separate directory
